- Update README.md for `autofill` usage. - Remove `autofill-no-op` component.
Update Autofill component:

- Create `AutofillInitializer` to automatically initialize the library.
- Remove AutofillService.Builder and split AutofillProcessor.
- Add logs to json read config.
- Add `matchAnyField` flag.
- Update `autofill` to `0.2.0`.
