--- libraries/autofill-service-no-op/.gitignore
@@ -1,1 +1,0 @@
-/build
