--- libraries/autofill-service-no-op/build.gradle
@@ -1,52 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-    id 'maven-publish'
-}
-
-android {
-    compileSdkVersion 33
-    buildToolsVersion "30.0.3"
-
-    defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion 33
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-
-        debug {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_11
-        targetCompatibility JavaVersion.VERSION_11
-    }
-    kotlinOptions {
-        jvmTarget = '11'
-    }
-    namespace 'com.trendyol.android.devtools.autofillservice'
-}
-
-ext {
-    PUBLISH_GROUP_ID = 'com.trendyol.android.devtools'
-    PUBLISH_VERSION = '0.1.0'
-    PUBLISH_ARTIFACT_ID = 'autofill-service-no-op'
-    PUBLISH_DESCRIPTION = "Android QA Form Autofill Service"
-    PUBLISH_URL = "https://github.com/Trendyol/android-dev-tools"
-    PUBLISH_LICENSE_NAME = "Android DevTools License"
-    PUBLISH_LICENSE_URL = "https://github.com/Trendyol/android-dev-tools/blob/master/LICENSE"
-    PUBLISH_SCM_CONNECTION = "scm:git:github.com/Trendyol/android-dev-tools.git"
-    PUBLISH_SCM_DEV_CONNECTION = "scm:git:ssh://github.com/Trendyol/android-dev-tools.git"
-    PUBLISH_SCM_URL = "https://github.com/Trendyol/android-dev-tools/tree/main"
-}
-
-apply from: "${rootProject.projectDir}/scripts/publish-module.gradle"
