--- libraries/autofill-service-no-op/src/main/java/com/trendyol/android/devtools/autofillservice/AutofillService.kt
@@ -1,16 +1,0 @@
-package com.trendyol.android.devtools.autofillservice
-
-import android.app.Application
-
-class AutofillService {
-
-    class Builder(private val application: Application) {
-
-        fun withFilePath(filePath: String): Builder {
-            return this
-        }
-
-        fun build() {
-        }
-    }
-}
