+++ libraries/autofill-service/build.gradle
@@ -35,6 +35,7 @@ android {
         jvmTarget = '11'
     }
     buildFeatures {
+        buildConfig false
         viewBinding true
     }
     namespace 'com.trendyol.android.devtools.autofillservice'
@@ -42,7 +43,7 @@ android {
 
 ext {
     PUBLISH_GROUP_ID = 'com.trendyol.android.devtools'
-    PUBLISH_VERSION = '0.1.0'
+    PUBLISH_VERSION = '0.2.0'
     PUBLISH_ARTIFACT_ID = 'autofill-service'
     PUBLISH_DESCRIPTION = "Android QA Form Autofill Service"
     PUBLISH_URL = "https://github.com/Trendyol/android-dev-tools"
@@ -57,6 +58,7 @@ apply from: "${rootProject.projectDir}/scripts/publish-module.gradle"
 
 dependencies {
     implementation "androidx.core:core-ktx:$ktx_version"
+    implementation "androidx.startup:startup-runtime:1.2.0"
     implementation "androidx.appcompat:appcompat:$appcompat_version"
     implementation "com.google.android.material:material:$material_version"
