+++ libraries/autofill-service/src/main/AndroidManifest.xml
@@ -1,0 +1,16 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <application>
+
+        <provider
+            android:name="androidx.startup.InitializationProvider"
+            android:authorities="${applicationId}.androidx-startup"
+            tools:node="merge">
+
+            <meta-data
+                android:name="com.trendyol.android.devtools.autofillservice.AutofillInitializer"
+                android:value="androidx.startup" />
+        </provider>
+    </application>
+</manifest>
