+++ libraries/autofill-service/src/main/java/com/trendyol/android/devtools/autofillservice/internal/model/FillData.kt
@@ -2,7 +2,7 @@ package com.trendyol.android.devtools.autofillservice.internal.model
 
 import com.squareup.moshi.Json
 
-data class FillData(
+internal data class FillData(
     @Json(name = "description") val description: String,
     @Json(name = "values") val values: List<String>,
 )
