+++ libraries/autofill-service/src/main/java/com/trendyol/android/devtools/autofillservice/internal/model/Form.kt
@@ -5,4 +5,5 @@ import com.squareup.moshi.Json
 internal data class Form(
     @Json(name = "fields") val fields: List<String>,
     @Json(name = "categories") val categories: Map<String, List<FillData>>,
+    @Json(name = "matchAnyField") val matchAnyField: Boolean = false,
 )
